# [IMP] mrp: use prefetch for recursive quantities computation