**Description of the issue/feature this PR addresses:**

1) Avoid calling super() on an empty recordset, in case all products are kits. Even on an empty recordset, some queries would be fired. This PR avoids them.

2) Use prefetch to read kit components quantity fields, to make use of the recorset optimizations that ends up saving sql queries. Otherwise, component quantities were computed on-demand one by one.

**Current behavior before PR:**

In case of kit products, whose components weren't included in `self`, their quantity fields were [read (thus, computed) on demand, one by one](https://github.com/odoo/odoo/blob/0f9f8409b379e3a0af28cecd3ed26829b9b26032/addons/mrp/models/product.py#L161-L171)

This is highly inefficient because each record computation will fire up several sql queries to read quants, and stock moves.

**Desired behavior after PR is merged:**

All non-cached components are identified beforehand, in order to leverage prefetching.

**Metrics**

On a production database with 3866 products, 855 of which are saleable and many are kits with recursive boms.

```py
from timeit import timeit
products = env['product.product'].search([('sale_ok', '=', True)])  # 855 products
timeit(lambda: products.read(['virtual_available']) and env.cache.invalidate(), number=100)
```

```
before: 234s (2340ms / iteration)
after: 68s (680ms / iteration)
```

and also the case of a single kit product composed of a 2 level recursive bom

```py
from timeit import timeit
products = env['product.product'].browse(10893)
timeit(lambda: products.read(['virtual_available']) and env.cache.invalidate(), number=100)
```

```
before: 4.64s (46ms / iteration)
after: 2.40s (24ms / iteration)
```




---

**NOTE**: In case of kits with recursive boms (kits with kits as components), the result is still not perfect but much better than before. We still end up in a recursive computation scenario, but at least now on each iteration all non-cached components are computed together, at once, leveraging the recorset optimizations (`read_group`, etc..)

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr